+++ tests/integration/instruments/test_pymongo.py
@@ -7,7 +7,7 @@ import os
 import pymongo
 import pytest
 
-from scout_apm.instruments.pymongo import ensure_installed
+from scout_apm.instruments.pymongo import COLLECTION_METHODS, ensure_installed
 from tests.compat import mock
 
 
@@ -18,6 +18,39 @@ def pymongo_client():
     if "MONGODB_URL" not in os.environ:
         raise pytest.skip("MongoDB isn't available")
     yield pymongo.MongoClient(os.environ["MONGODB_URL"])
+
+
+def test_all_collection_attributes_accounted_for():
+    all_methods = {
+        m for m in dir(pymongo.collection.Collection) if not m.startswith("_")
+    }
+    deliberately_ignored_methods = {
+        # Properties:
+        "codec_options",
+        "database",
+        "full_name",
+        "name",
+        "read_concern",
+        "read_preference",
+        "write_concern",
+        # Non-querying methods:
+        "initialize_ordered_bulk_op",
+        "initialize_unordered_bulk_op",
+        "next",
+        "options",
+        "with_options",
+        # Returns a long running iterator rather than doing a direct query;
+        # probably won't be used in web requests:
+        "watch",
+    }
+    assert (
+        all_methods - deliberately_ignored_methods - set(COLLECTION_METHODS)
+    ) == set()
+
+
+@pytest.mark.parametrize(["method_name"], [[x] for x in COLLECTION_METHODS])
+def test_all_collection_methods_exist(method_name):
+    assert hasattr(pymongo.collection.Collection, method_name)
 
 
 def test_ensure_installed_twice(caplog):